Rename Batch as TransactionalBatch (#965)
* Rename public name for batch as TransactionalBatch

* Remove batch from preview

* Update contract file

* Add to changelog
