+++ Microsoft.Azure.Cosmos/src/Batch/PartitionKeyRangeBatchResponse.cs
@@ -13,11 +13,11 @@ namespace Microsoft.Azure.Cosmos
     /// <summary>
     /// Response of a cross partition key batch request.
     /// </summary>
-    internal class PartitionKeyRangeBatchResponse : BatchResponse
+    internal class PartitionKeyRangeBatchResponse : TransactionalBatchResponse
     {
         // Results sorted in the order operations had been added.
-        private readonly BatchOperationResult[] resultsByOperationIndex;
-        private readonly BatchResponse serverResponse;
+        private readonly TransactionalBatchOperationResult[] resultsByOperationIndex;
+        private readonly TransactionalBatchResponse serverResponse;
         private bool isDisposed;
 
         /// <summary>
@@ -28,13 +28,13 @@ namespace Microsoft.Azure.Cosmos
         /// <param name="serializer">Serializer to deserialize response resource body streams.</param>
         internal PartitionKeyRangeBatchResponse(
             int originalOperationsCount,
-            BatchResponse serverResponse,
+            TransactionalBatchResponse serverResponse,
             CosmosSerializer serializer)
         {
             this.StatusCode = serverResponse.StatusCode;
 
             this.serverResponse = serverResponse;
-            this.resultsByOperationIndex = new BatchOperationResult[originalOperationsCount];
+            this.resultsByOperationIndex = new TransactionalBatchOperationResult[originalOperationsCount];
 
             StringBuilder errorMessageBuilder = new StringBuilder();
             List<ItemBatchOperation> itemBatchOperations = new List<ItemBatchOperation>();
@@ -78,7 +78,7 @@ namespace Microsoft.Azure.Cosmos
         public override int Count => this.resultsByOperationIndex.Length;
 
         /// <inheritdoc />
-        public override BatchOperationResult this[int index] => this.resultsByOperationIndex[index];
+        public override TransactionalBatchOperationResult this[int index] => this.resultsByOperationIndex[index];
 
         /// <summary>
         /// Gets the result of the operation at the provided index in the batch - the returned result has a Resource of provided type.
@@ -86,14 +86,14 @@ namespace Microsoft.Azure.Cosmos
         /// <typeparam name="T">Type to which the Resource in the operation result needs to be deserialized to, when present.</typeparam>
         /// <param name="index">0-based index of the operation in the batch whose result needs to be returned.</param>
         /// <returns>Result of batch operation that contains a Resource deserialized to specified type.</returns>
-        public override BatchOperationResult<T> GetOperationResultAtIndex<T>(int index)
+        public override TransactionalBatchOperationResult<T> GetOperationResultAtIndex<T>(int index)
         {
             if (index >= this.Count)
             {
                 throw new IndexOutOfRangeException();
             }
 
-            BatchOperationResult result = this.resultsByOperationIndex[index];
+            TransactionalBatchOperationResult result = this.resultsByOperationIndex[index];
 
             T resource = default(T);
             if (result.ResourceStream != null)
@@ -101,16 +101,16 @@ namespace Microsoft.Azure.Cosmos
                 resource = this.Serializer.FromStream<T>(result.ResourceStream);
             }
 
-            return new BatchOperationResult<T>(result, resource);
+            return new TransactionalBatchOperationResult<T>(result, resource);
         }
 
         /// <summary>
         /// Gets an enumerator over the operation results.
         /// </summary>
         /// <returns>Enumerator over the operation results.</returns>
-        public override IEnumerator<BatchOperationResult> GetEnumerator()
+        public override IEnumerator<TransactionalBatchOperationResult> GetEnumerator()
         {
-            foreach (BatchOperationResult result in this.resultsByOperationIndex)
+            foreach (TransactionalBatchOperationResult result in this.resultsByOperationIndex)
             {
                 yield return result;
             }
