+++ Microsoft.Azure.Cosmos/src/Batch/TransactionalBatch.cs
@@ -11,23 +11,18 @@ namespace Microsoft.Azure.Cosmos
     /// <summary>
     /// Represents a batch of requests that will be performed atomically against the Azure Cosmos DB service.
     /// </summary>
-#if PREVIEW
-    public
-#else
-    internal
-#endif
-    abstract class Batch
+    public abstract class TransactionalBatch
     {
         /// <summary>
         /// Adds an operation to create an item into the batch.
         /// </summary>
         /// <param name="item">A JSON serializable object that must contain an id property.<see cref="CosmosSerializer"/> to implement a custom serializer.</param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
         /// <typeparam name="T">The type of item to be created.</typeparam>
-        public abstract Batch CreateItem<T>(
+        public abstract TransactionalBatch CreateItem<T>(
             T item,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to create an item into the batch.
@@ -36,32 +31,32 @@ namespace Microsoft.Azure.Cosmos
         /// A <see cref="Stream"/> containing the payload of the item.
         /// The stream must have a UTF-8 encoded JSON object which contains an id property.
         /// </param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
-        public abstract Batch CreateItemStream(
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
+        public abstract TransactionalBatch CreateItemStream(
             Stream streamPayload,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to read an item into the batch.
         /// </summary>
         /// <param name="id">The cosmos item id.</param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
-        public abstract Batch ReadItem(
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
+        public abstract TransactionalBatch ReadItem(
             string id,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to upsert an item into the batch.
         /// </summary>
         /// <param name="item">A JSON serializable object that must contain an id property. <see cref="CosmosSerializer"/> to implement a custom serializer.</param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
         /// <typeparam name="T">The type of item to be created.</typeparam>
-        public abstract Batch UpsertItem<T>(
+        public abstract TransactionalBatch UpsertItem<T>(
             T item,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to upsert an item into the batch.
@@ -70,24 +65,24 @@ namespace Microsoft.Azure.Cosmos
         /// A <see cref="Stream"/> containing the payload of the item.
         /// The stream must have a UTF-8 encoded JSON object which contains an id property.
         /// </param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
-        public abstract Batch UpsertItemStream(
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
+        public abstract TransactionalBatch UpsertItemStream(
             Stream streamPayload,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to replace an item into the batch.
         /// </summary>
         /// <param name="id">The cosmos item id.</param>
         /// <param name="item">A JSON serializable object that must contain an id property. <see cref="CosmosSerializer"/> to implement a custom serializer.</param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
         /// <typeparam name="T">The type of item to be created.</typeparam>
-        public abstract Batch ReplaceItem<T>(
+        public abstract TransactionalBatch ReplaceItem<T>(
             string id,
             T item,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to replace an item into the batch.
@@ -97,29 +92,29 @@ namespace Microsoft.Azure.Cosmos
         /// A <see cref="Stream"/> containing the payload of the item.
         /// The stream must have a UTF-8 encoded JSON object which contains an id property.
         /// </param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
-        public abstract Batch ReplaceItemStream(
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
+        public abstract TransactionalBatch ReplaceItemStream(
             string id,
             Stream streamPayload,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Adds an operation to delete an item into the batch.
         /// </summary>
         /// <param name="id">The cosmos item id.</param>
-        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="BatchItemRequestOptions"/>.</param>
-        /// <returns>The <see cref="Batch"/> instance with the operation added.</returns>
-        public abstract Batch DeleteItem(
+        /// <param name="requestOptions">(Optional) The options for the item request. <see cref="TransactionalBatchItemRequestOptions"/>.</param>
+        /// <returns>The <see cref="TransactionalBatch"/> instance with the operation added.</returns>
+        public abstract TransactionalBatch DeleteItem(
             string id,
-            BatchItemRequestOptions requestOptions = null);
+            TransactionalBatchItemRequestOptions requestOptions = null);
 
         /// <summary>
         /// Executes the batch at the Azure Cosmos service as an asynchronous operation.
         /// </summary>
         /// <param name="cancellationToken">(Optional) <see cref="CancellationToken"/> representing request cancellation.</param>
-        /// <returns>An awaitable <see cref="BatchResponse"/> which contains the completion status and results of each operation.</returns>
-        public abstract Task<BatchResponse> ExecuteAsync(
+        /// <returns>An awaitable <see cref="TransactionalBatchResponse"/> which contains the completion status and results of each operation.</returns>
+        public abstract Task<TransactionalBatchResponse> ExecuteAsync(
             CancellationToken cancellationToken = default(CancellationToken));
     }
 }
