+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/Batch/BatchTestBase.cs
@@ -342,9 +342,9 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
             return doc.Id;
         }
 
-        internal static BatchItemRequestOptions GetBatchItemRequestOptions(TestDoc doc, bool isSchematized, bool useEpk = false, int? ttlInSeconds = null)
+        internal static TransactionalBatchItemRequestOptions GetBatchItemRequestOptions(TestDoc doc, bool isSchematized, bool useEpk = false, int? ttlInSeconds = null)
         {
-            BatchItemRequestOptions requestOptions = new BatchItemRequestOptions()
+            TransactionalBatchItemRequestOptions requestOptions = new TransactionalBatchItemRequestOptions()
             {
                 Properties = new Dictionary<string, object>()
             };
