+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Batch/BatchAsyncContainerExecutorTests.cs
@@ -58,7 +58,7 @@ namespace Microsoft.Azure.Cosmos.Tests
                 string.Empty);
             mockContainer.Setup(x => x.GetRoutingMapAsync(It.IsAny<CancellationToken>())).Returns(Task.FromResult(routingMap));
             BatchAsyncContainerExecutor executor = new BatchAsyncContainerExecutor(mockContainer.Object, mockedContext.Object, 20, Constants.MaxDirectModeBatchRequestBodySizeInBytes, 1);
-            BatchOperationResult result = await executor.AddAsync(itemBatchOperation);
+            TransactionalBatchOperationResult result = await executor.AddAsync(itemBatchOperation);
 
             Mock.Get(mockContainer.Object)
                 .Verify(x => x.GetPartitionKeyDefinitionAsync(It.IsAny<CancellationToken>()), Times.Exactly(2));
@@ -124,7 +124,7 @@ namespace Microsoft.Azure.Cosmos.Tests
                 string.Empty);
             mockContainer.Setup(x => x.GetRoutingMapAsync(It.IsAny<CancellationToken>())).Returns(Task.FromResult(routingMap));
             BatchAsyncContainerExecutor executor = new BatchAsyncContainerExecutor(mockContainer.Object, mockedContext.Object, 20, Constants.MaxDirectModeBatchRequestBodySizeInBytes, 1);
-            BatchOperationResult result = await executor.AddAsync(itemBatchOperation);
+            TransactionalBatchOperationResult result = await executor.AddAsync(itemBatchOperation);
 
             Mock.Get(mockContainer.Object)
                 .Verify(x => x.GetPartitionKeyDefinitionAsync(It.IsAny<CancellationToken>()), Times.Exactly(2));
@@ -190,7 +190,7 @@ namespace Microsoft.Azure.Cosmos.Tests
                 string.Empty);
             mockContainer.Setup(x => x.GetRoutingMapAsync(It.IsAny<CancellationToken>())).Returns(Task.FromResult(routingMap));
             BatchAsyncContainerExecutor executor = new BatchAsyncContainerExecutor(mockContainer.Object, mockedContext.Object, 20, Constants.MaxDirectModeBatchRequestBodySizeInBytes, 1);
-            BatchOperationResult result = await executor.AddAsync(itemBatchOperation);
+            TransactionalBatchOperationResult result = await executor.AddAsync(itemBatchOperation);
 
             Mock.Get(mockContainer.Object)
                 .Verify(x => x.GetPartitionKeyDefinitionAsync(It.IsAny<CancellationToken>()), Times.Exactly(2));
@@ -255,7 +255,7 @@ namespace Microsoft.Azure.Cosmos.Tests
                 string.Empty);
             mockContainer.Setup(x => x.GetRoutingMapAsync(It.IsAny<CancellationToken>())).Returns(Task.FromResult(routingMap));
             BatchAsyncContainerExecutor executor = new BatchAsyncContainerExecutor(mockContainer.Object, mockedContext.Object, 20, Constants.MaxDirectModeBatchRequestBodySizeInBytes, 1);
-            BatchOperationResult result = await executor.AddAsync(itemBatchOperation);
+            TransactionalBatchOperationResult result = await executor.AddAsync(itemBatchOperation);
 
             Mock.Get(mockContainer.Object)
                 .Verify(x => x.GetPartitionKeyDefinitionAsync(It.IsAny<CancellationToken>()), Times.Once);
@@ -275,10 +275,10 @@ namespace Microsoft.Azure.Cosmos.Tests
 
         private async Task<ResponseMessage> GenerateSplitResponseAsync(ItemBatchOperation itemBatchOperation)
         {
-            List<BatchOperationResult> results = new List<BatchOperationResult>();
+            List<TransactionalBatchOperationResult> results = new List<TransactionalBatchOperationResult>();
             ItemBatchOperation[] arrayOperations = new ItemBatchOperation[1];
             results.Add(
-                new BatchOperationResult(HttpStatusCode.Gone)
+                new TransactionalBatchOperationResult(HttpStatusCode.Gone)
                 {
                     ETag = itemBatchOperation.Id,
                     SubStatusCode = SubStatusCodes.PartitionKeyRangeGone
@@ -313,10 +313,10 @@ namespace Microsoft.Azure.Cosmos.Tests
 
         private async Task<ResponseMessage> GenerateCacheStaleResponseAsync(ItemBatchOperation itemBatchOperation)
         {
-            List<BatchOperationResult> results = new List<BatchOperationResult>();
+            List<TransactionalBatchOperationResult> results = new List<TransactionalBatchOperationResult>();
             ItemBatchOperation[] arrayOperations = new ItemBatchOperation[1];
             results.Add(
-                new BatchOperationResult(HttpStatusCode.Gone)
+                new TransactionalBatchOperationResult(HttpStatusCode.Gone)
                 {
                     ETag = itemBatchOperation.Id,
                     SubStatusCode = SubStatusCodes.NameCacheIsStale
@@ -351,10 +351,10 @@ namespace Microsoft.Azure.Cosmos.Tests
 
         private async Task<ResponseMessage> Generate429ResponseAsync(ItemBatchOperation itemBatchOperation)
         {
-            List<BatchOperationResult> results = new List<BatchOperationResult>();
+            List<TransactionalBatchOperationResult> results = new List<TransactionalBatchOperationResult>();
             ItemBatchOperation[] arrayOperations = new ItemBatchOperation[1];
             results.Add(
-                new BatchOperationResult((HttpStatusCode) StatusCodes.TooManyRequests)
+                new TransactionalBatchOperationResult((HttpStatusCode) StatusCodes.TooManyRequests)
                 {
                     ETag = itemBatchOperation.Id
                 });
@@ -387,10 +387,10 @@ namespace Microsoft.Azure.Cosmos.Tests
 
         private async Task<ResponseMessage> GenerateOkResponseAsync(ItemBatchOperation itemBatchOperation)
         {
-            List<BatchOperationResult> results = new List<BatchOperationResult>();
+            List<TransactionalBatchOperationResult> results = new List<TransactionalBatchOperationResult>();
             ItemBatchOperation[] arrayOperations = new ItemBatchOperation[1];
             results.Add(
-                new BatchOperationResult(HttpStatusCode.OK)
+                new TransactionalBatchOperationResult(HttpStatusCode.OK)
                 {
                     ETag = itemBatchOperation.Id
                 });
